requirements.txt
asgiref==3.3.4
Django==3.2
djangorestframework==3.12.4
pytz==2021.1
sqlparse==0.4.1

requests~=2.25.1
bs4~=0.0.1
beautifulsoup4~=4.9.3